[Feature] security 적용, AOP 로깅 개인정보 마스킹 및 VO 타입 정상화 반영 (#3)
* feat: DB 연동 커스텀 시큐리티 인증 설정 및 매퍼 구현

* chore: 사용하지 않는 mappers/.gitkeep 삭제

* feat: 시큐리티 DB 인증 추가 및 MyBatis Interceptor/테스트 코드 반영

* refactor: DDL 규격에 맞추어 모든 VO 클래스 데이터 타입 재정비

* security: AOP 로깅 개인정보 마스킹 및 VO 타입 정상화 반영

* refactor: 감사 필드 비인증 시 예외 발생 처리 및 공통 에러 관리 구조 구현

* refactor: UserMapper.xml 쿼리 고도화 및 CustomUserDetailsService 로깅 보안 취약점 패치

* fix: insertAuthority 파라미터 AuthoritiesVO로 변경하여 MyBatis 바인딩 및 인터셉터 호환 에러 해결

+++ src/main/java/com/example/smartmessaging/dto/vo/AuthoritiesVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,6 +12,6 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class AuthoritiesVO extends BaseVO {
-    private BigDecimal userId;
+    private Long userId;
     private String authority;
 }
