+++ src/main/java/com/example/smartmessaging/dto/vo/CustomerTagMappingVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,6 +12,6 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class CustomerTagMappingVO extends BaseVO {
-    private BigDecimal tagId;
-    private BigDecimal customerId;
+    private Long tagId;
+    private Long customerId;
 }
