+++ src/main/java/com/example/smartmessaging/dto/vo/RejectHistoryVO.java
@@ -5,7 +5,6 @@ import lombok.Setter;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
-import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,7 +12,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class RejectHistoryVO extends BaseVO {
-    private BigDecimal id;
-    private BigDecimal customerId;
+    private Long id;
+    private Long customerId;
     private Boolean isRejected;
 }
